Config implementation file added
Because it feels to weird to put everything in the .h file. Don't really
know what is the standard way to do things in that front. The .cpp file
is a little bit more trouble because the Makefile has to be updated (by
the way, not sure if I'm doing the Makefile thing right either...).