Added bin directory to the repo.
Not very nice to have the executables in the repo, but this way git
keeps track of the bin/ folder. Wihtout the bin/ folder, the `make`
command fails (can't create output file).

I know, it is very easy to fix the Makefile to recreate the dir,
but... that's a task for another day.